@@ -1,9 +1,5 @@
 module webapp
 
-go 1.14
+go 1.20
 
-require (
-	github.com/gorilla/mux v1.7.4
-	github.com/gorilla/securecookie v1.1.1
-	github.com/joho/godotenv v1.3.0
-)
+require github.com/gorilla/mux v1.8.0 // indirect
