@@ -4,18 +4,12 @@ import (
 	"fmt"
 	"log"
 	"net/http"
-	"webapp/src/config"
-	"webapp/src/cookies"
 	"webapp/src/router"
-	"webapp/src/utils"
 )
 
-func main() {
-	config.Carregar()
-	cookies.Configurar()
-	utils.CarregarTemplates()
-	r := router.Gerar()
+func main(){
+	fmt.Println("Rodando WebApp!")
 
-	fmt.Printf("Escutando na porta %d\n", config.Porta)
-	log.Fatal(http.ListenAndServe(fmt.Sprintf(":%d", config.Porta), r))
+	r := router.Gerar()
+	log.Fatal(http.ListenAndServe(":3000", r))
 }
